+++ Changelog.md
@@ -1,6 +1,8 @@
 # Change log
 
 ## master (unreleased)
+### New features
+* New method `BugzillaHelper#bug_exists?`
 
 ## 0.1.1 (2018-05-15)
 * Fix getting id from string with spaces